requirements.txt
# Python 3.10 | PyTorch 2.1.0 | CUDA 12.1
# uv pip install -r requirements.txt

torch==2.1.0+cu121
torchvision==0.16.0+cu121
torch-geometric==2.4.0
torch-scatter
torch-sparse
torch-cluster
networkx>=3.1
numpy>=1.24,<2
tqdm>=4.65
pyyaml>=6.0
scikit-learn>=1.3
tensorboard>=2.14
